requirements.txt
alembic==1.16.5
APScheduler==3.11.2
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.1.8
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.48
tomli==2.4.0
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
Werkzeug==3.1.3
zipp==3.23.0
gunicorn==23.0.0
